Extract active sun session module
