+++ js/sun.js
@@ -27,8 +27,25 @@ import {
   _testRegionBandLandmarks,
 } from './sun-body-silhouette.js';
 import {
+  configureSunActiveSession,
+  quickLogSunSession,
+  openStartSunSessionDialog,
+  _wireBackdropClose,
+  trapModalFocus,
+  _formatElapsed,
+  liveDosesFor as _liveDosesFor,
+  commitSunLiveSlice as _commitCurrentSlice,
+  setSunLiveState as _setLiveState,
+  clearSunLiveState as _clearLiveState,
+  ensureActiveTicker as _ensureActiveTicker,
+  resumeActiveTickerIfNeeded as _resumeActiveTickerIfNeeded,
+  hydrateSunSessionFromProfileCoords as _hydrateFromProfileCoords,
+  resetSunActiveSessionState,
+  POSTURE_MULTIPLIERS,
+  SURFACE_ALBEDO,
+} from './sun-active-session.js';
+import {
   configureSunSessionUI,
-  renderChannelChips,
   renderSessionsList,
   renderSunSessionRow,
   openDetailedSessionDialog,
@@ -38,6 +55,7 @@ import {
 } from './sun-session-ui.js';
 export { BODY_REGIONS, renderBodySilhouette, bindBodySilhouette };
 export { renderSessionsList, renderSunSessionRow, openDetailedSessionDialog, openSunSessionDetail };
+export { quickLogSunSession, openStartSunSessionDialog, _wireBackdropClose, trapModalFocus };
 // NOTE: sun-ai-analysis.js is intentionally NOT imported here — it
 // imports from this file (getSessions, formatChannelUnit, etc.), and a
 // reciprocal import would create a circular dependency that risks TDZ
@@ -309,7 +327,7 @@ export function getActiveSession() {
 }
 
 // Posture options surfaced in pickers + applied as a multiplier on the
-// effective body fraction (see _POSTURE_MULTIPLIERS in _rateAtInstant).
+// effective body fraction (see POSTURE_MULTIPLIERS in sun-active-session.js).
 export const POSTURE_OPTIONS = [
   { key: 'standing',     label: 'Standing / walking' },
   { key: 'sitting',      label: 'Sitting / reclined' },
@@ -318,7 +336,7 @@ export const POSTURE_OPTIONS = [
 ];
 
 // Surface albedo dropdown values — UV reflection from below augments
-// total received irradiance by ~(albedo × 0.5). See _SURFACE_ALBEDO.
+// total received irradiance by ~(albedo × 0.5). See SURFACE_ALBEDO.
 export const SURFACE_OPTIONS = [
   { key: 'grass',    label: 'Grass / dirt (~3% reflect)' },
   { key: 'concrete', label: 'Concrete / pavement (~10%)' },
@@ -365,8 +383,8 @@ export async function startSession({ exposurePreset = 'face_hands', regions, eye
     // 🔄 Flip button (calls flipSidesMidSession).
     bodyExposure: { preset: preset.key, fraction, regions: regionsArr, sunscreenSPF: null, glassBetween, rotatedSides: !!rotatedSides },
     eyeExposure: { mode: eyeMode, lensTint, durationSec: null }, // durationSec assigned at stop
-    posture,                  // body orientation multiplier — see _POSTURE_MULTIPLIERS
-    surfaceAlbedo,            // ground reflectance multiplier — see _SURFACE_ALBEDO
+    posture,                  // body orientation multiplier — see POSTURE_MULTIPLIERS
+    surfaceAlbedo,            // ground reflectance multiplier — see SURFACE_ALBEDO
     atmosphere: null, // populated at stop or fetched async
     doses: null,
     safety: null,
@@ -838,8 +856,8 @@ export async function hydrateSession(id, { lat, lon } = {}) {
     // Apply posture + surface-albedo multipliers to body fraction so
     // hydrated doses match the live engine's accounting.
     const baseFraction = sess.bodyExposure?.fraction ?? 0;
-    const postureMult = _POSTURE_MULTIPLIERS[sess.posture] ?? 1.0;
-    const albedoMult = 1 + (_SURFACE_ALBEDO[sess.surfaceAlbedo] ?? 0) * 0.5;
+    const postureMult = POSTURE_MULTIPLIERS[sess.posture] ?? 1.0;
+    const albedoMult = 1 + (SURFACE_ALBEDO[sess.surfaceAlbedo] ?? 0) * 0.5;
     const effFraction = baseFraction * postureMult * albedoMult;
     sess.doses = computeChannelDoses({
       spectrum,
@@ -1318,407 +1336,6 @@ export function cumulativeMEDYesterday() {
   return total;
 }
 
-// ─── UI: Quick log ─────────────────────────────────────────────────────
-
-// Single-tap "I'm outside now" — starts a session with last-used defaults.
-// On stop: skips confirm dialog (user explicitly tapped stop). Notification
-// includes duration + the channel that benefited most for instant feedback.
-export async function quickLogSunSession() {
-  const active = getActiveSession();
-  if (active) {
-    await stopSession(active.id);
-    await _hydrateFromProfileCoords(active.id);
-    const sess = getSessions().find(s => s.id === active.id);
-    const dur = Math.round(sess?.durationMin || 0);
-    const summary = _plainStopSummary(sess, dur);
-    showNotification(summary, summary.includes('over your burn threshold') ? 'error' : 'success', 7000);
-    _refreshSurfaces();
-    return;
-  }
-  // No active session — open the silhouette picker so the user can pick
-  // exposed regions before the session begins. Inherits from last session.
-  return openStartSunSessionDialog();
-}
-
-// Lookup current UVI from the configured atm provider. Returns the scalar
-// uvIndex or null on any failure (no coords / fetch error / missing field).
-// Used for the pre-session high-UV warning banner.
-async function _fetchCurrentUVI() {
-  if (!window.fetchAtmosphere) return null;
-  const coords = getSunCoords();
-  if (!coords) return null;
-  try {
-    const atm = await window.fetchAtmosphere({
-      lat: coords.lat, lon: coords.lon, isoTime: new Date().toISOString(),
-    });
-    const overridden = _applyAtmOverrides(atm);
-    return overridden?.uvIndex ?? null;
-  } catch (e) { return null; }
-}
-
-// Estimated minutes-to-MED for a given UVI + Fitzpatrick + photosensitive
-// status. The CIE-erythemal action spectrum + UVI definition give us
-// MED time = baseMED_J_per_m2 / (UVI × ~25 mW/m²). Photosensitive meds
-// scale the MED denominator down via PHOTOSENSITIVE_MED_TIERS.
-function _estimateMedMinutes(uvi, fitzpatrick, psmTier) {
-  if (!Number.isFinite(uvi) || uvi <= 0) return null;
-  const fitzMED = { I: 200, II: 250, III: 300, IV: 450, V: 600, VI: 1000 };
-  const baseMED = fitzMED[fitzpatrick] ?? fitzMED.III;
-  const med = baseMED * (photosensitiveMedScale(psmTier) || 1.0);
-  // 1 UVI unit = 25 mW/m² CIE-erythemal-weighted irradiance.
-  const irradiance = uvi * 25; // mW/m²
-  const seconds = (med * 1000) / irradiance; // J/m² ÷ mW/m² → seconds (×1000 for unit alignment)
-  return Math.round(seconds / 60);
-}
-
-// Render the pre-session UVI banner HTML. Returns '' when conditions
-// don't warrant a warning (UVI < 8 OR Fitz IV-VI without photosensitive
-// meds). Always shows when photosensitiveMeds is moderate/severe even
-// at lower UVI because their MED is sharply lowered.
-function _renderUVIPreflightBanner(uvi, fitzpatrick, psmTier) {
-  if (!Number.isFinite(uvi)) return '';
-  const fairSkin = ['I', 'II', 'III'].includes(fitzpatrick);
-  const psmHigh = psmTier === 'moderate' || psmTier === 'severe';
-  // Don't pester at low UVI for non-fair, non-photosensitive users.
-  if (uvi < 8 && !psmHigh) return '';
-  if (uvi < 5 && !psmHigh) return '';
-  const medMin = _estimateMedMinutes(uvi, fitzpatrick, psmTier);
-  let cls = 'sun-uvi-warn';
-  let icon = '☀';
-  let title = '';
-  if (uvi >= 11) { cls = 'sun-uvi-extreme'; icon = '⚠'; title = `Extreme UV (UVI ${uvi.toFixed(1)})`; }
-  else if (uvi >= 8) { cls = 'sun-uvi-veryhigh'; icon = '☀'; title = `Very high UV (UVI ${uvi.toFixed(1)})`; }
-  else { title = `UV ${uvi.toFixed(1)} — burn risk elevated by photosensitizer`; }
-  const medLine = medMin ? `Estimated MED for Fitzpatrick ${fitzpatrick}${psmHigh ? ` + ${psmTier} photosensitizer` : ''}: ~${medMin} min uncovered.` : '';
-  return `<div class="${cls}"><strong>${icon} ${escapeHTML(title)}</strong> ${escapeHTML(medLine)} Sunscreen + cover up + a shorter session strongly suggested.</div>`;
-}
-
-function _buildStartSessionToast({ regionCount, uvi, psmTier, eyeMode }) {
-  const parts = [`Outdoor session started · ${regionCount} region${regionCount === 1 ? '' : 's'} exposed`];
-  const notes = [];
-  if (Number.isFinite(uvi) && uvi >= 11) notes.push(`extreme UV ${uvi.toFixed(1)}`);
-  else if (Number.isFinite(uvi) && uvi >= 8) notes.push(`high UV ${uvi.toFixed(1)}`);
-  const tier = _normalizePSMTier(psmTier);
-  if (tier !== 'none') notes.push(`${tier} photosensitizer`);
-  if (eyeMode === 'direct') notes.push('eyes uncovered');
-  if (notes.length) parts.push(`${notes.join(' + ')} · keep it short`);
-  return parts.join(' · ');
-}
-
-// Show the "What's uncovered?" dialog with the body silhouette + a Start
-// button. The picker pre-selects regions from the user's last completed
-// session so habitual users hit Start without changes; first-time users
-// pick everything fresh.
-//
-// Pre-flight UVI warning: when current UVI is in the high range (≥8) and
-// the user is a fair skin type (Fitzpatrick I-III), prepend an alert
-// banner with the estimated MED time for plain-text comprehension.
-export async function openStartSunSessionDialog() {
-  const last = getSessions().filter(s => s.endedAt).slice(-1)[0];
-  const lastRegions = new Set(last?.bodyExposure?.regions || []);
-  const defaultEye = last?.eyeExposure?.mode || 'direct';
-  const defaultLens = last?.eyeExposure?.lensTint || 'clear';
-  const defaultGlass = !!last?.bodyExposure?.glassBetween;
-  const defaultPosture = last?.posture || 'standing';
-  const defaultSurface = last?.surfaceAlbedo || 'grass';
-  // Pull current UVI for the high-UV pre-flight banner. Fire-and-forget;
-  // dialog opens immediately even if the fetch lags. Banner lights up
-  // when the promise resolves (slot in the modal).
-  const fitz = state.importedData?.sunDefaults?.fitzpatrick || 'III';
-  const psm = state.importedData?.sunDefaults?.photosensitiveMeds || 'none';
-  const uviPromise = _fetchCurrentUVI();
-  let latestPreflightUvi = null;
-
-  const overlay = document.createElement('div');
-  overlay.className = 'modal-overlay show';
-  overlay.innerHTML = `<div class="modal sun-start-modal" role="dialog" aria-label="Start sun session">
-    <div class="modal-header">
-      <h3>Start a sun session</h3>
-      <button class="modal-close" onclick="this.closest('.modal-overlay').remove()" aria-label="Close">×</button>
-    </div>
-    <div class="modal-body">
-      <div id="sun-start-uvi-banner" class="sun-start-uvi-banner" hidden></div>
-      <p class="modal-body-hint">Tap each body region that's uncovered right now. The session begins as soon as you hit Start.</p>
-      <div class="sun-silhouette-wrap" id="sun-start-silhouette-slot">${renderBodySilhouette(lastRegions)}</div>
-      <div class="sun-silhouette-hint-row" style="display:flex;align-items:center;justify-content:space-between;gap:8px">
-        <div class="sun-silhouette-hint" id="sun-start-hint">Tap any body region to toggle whether it's uncovered.</div>
-        <button type="button" class="ctx-btn-option" id="sun-start-clear" style="padding:2px 10px;font-size:11px">Clear</button>
-      </div>
-
-      <details class="sun-start-details">
-        <summary>Posture, surface, eyewear, sunscreen, glass — change defaults</summary>
-        <div class="sun-detailed-row" style="margin-top:10px">
-          <label class="ctx-label">Posture
-            <select id="start-posture" class="ctx-select">
-              ${POSTURE_OPTIONS.map(o => `<option value="${escapeAttr(o.key)}"${o.key === defaultPosture ? ' selected' : ''}>${escapeHTML(o.label)}</option>`).join('')}
-            </select>
-          </label>
-          <label class="ctx-label">Surface
-            <select id="start-surface" class="ctx-select">
-              ${SURFACE_OPTIONS.map(o => `<option value="${escapeAttr(o.key)}"${o.key === defaultSurface ? ' selected' : ''}>${escapeHTML(o.label)}</option>`).join('')}
-            </select>
-          </label>
-        </div>
-        <p class="sun-detailed-glass-hint">Lying flat catches more sun than standing (~40%). Reflective surfaces (sand, water, snow) bounce UV onto your skin from below.</p>
-        <div class="sun-detailed-row" style="margin-top:10px">
-          <label class="ctx-label">Eyes
-            <select id="start-eye-mode" class="ctx-select">
-              ${EYE_MODES.map(e => `<option value="${escapeAttr(e.key)}"${e.key === defaultEye ? ' selected' : ''}>${escapeHTML(e.pickerLabel || e.label)}</option>`).join('')}
-            </select>
-          </label>
-          <label class="ctx-label">Lens tint
-            <select id="start-lens-tint" class="ctx-select">
-              ${LENS_TINTS.map(l => `<option value="${escapeAttr(l.key)}"${l.key === defaultLens ? ' selected' : ''}>${escapeHTML(l.label)}</option>`).join('')}
-            </select>
-          </label>
-        </div>
-        <div class="ctx-label sun-detailed-glass" style="margin-top:8px;display:flex;align-items:center;justify-content:space-between;gap:12px">
-          <span style="flex:1;min-width:0">Behind glass (window / car / sunroom)</span>
-          <label class="toggle-switch">
-            <input type="checkbox" id="start-glass"${defaultGlass ? ' checked' : ''} />
-            <span class="toggle-slider"></span>
-          </label>
-        </div>
-        <p class="sun-detailed-glass-hint">Standard window glass blocks ~99% of UVB. Vitamin D synthesis stops; circadian and warmth signals still get through. We zero the burn dose accordingly. (Want to measure YOUR glass's transmission? Light tools → Window check.)</p>
-        <div class="ctx-label sun-detailed-glass" style="margin-top:8px;display:flex;align-items:center;justify-content:space-between;gap:12px">
-          <span style="flex:1;min-width:0">Plan to flip front ↔ back during the session</span>
-          <label class="toggle-switch">
-            <input type="checkbox" id="start-rotated" />
-            <span class="toggle-slider"></span>
-          </label>
-        </div>
-        <p class="sun-detailed-glass-hint">Toggle on if you'll alternate sides — doubles the vitamin D estimate to reflect that fresh skin keeps synthesizing after the first side approaches saturation. You can also tap 🔄 Flip mid-session.</p>
-      </details>
-
-      <div class="modal-actions" style="margin-top:18px">
-        <button class="import-btn import-btn-secondary" onclick="this.closest('.modal-overlay').remove()">Cancel</button>
-        <button class="import-btn import-btn-primary" id="start-confirm">☀ Start session</button>
-      </div>
-    </div>
-  </div>`;
-  _wireBackdropClose(overlay);
-  document.body.appendChild(overlay);
-  trapModalFocus(overlay);
-
-  const selected = new Set(lastRegions);
-  const slot = overlay.querySelector('#sun-start-silhouette-slot');
-  const hint = overlay.querySelector('#sun-start-hint');
-  const updateHint = () => {
-    const fraction = Array.from(selected).reduce((sum, key) => {
-      const r = BODY_REGIONS.find(b => b.key === key);
-      return sum + (r?.fraction || 0);
-    }, 0);
-    if (selected.size === 0) {
-      hint.textContent = 'Tap any body region to toggle whether it\'s uncovered.';
-    } else {
-      const labels = Array.from(selected).map(k => BODY_REGIONS.find(b => b.key === k)?.label || k).join(', ');
-      // See note in the speed-log handler above re: 0.95 ceiling.
-      const pctLabel = fraction >= 0.94 ? 'full body' : `${(fraction * 100).toFixed(0)}% of skin`;
-      hint.textContent = `${selected.size} region${selected.size === 1 ? '' : 's'} exposed (${pctLabel}) — ${labels}`;
-    }
-  };
-  // Clear — single bulk-deselect affordance. The silhouette itself is
-  // the picker; pre-2026-05-08 also had a 4-zone toggle row but it was
-  // a redundant abstraction. See light-devices.js openDeviceSessionDialog
-  // for the matching change on the PBM side.
-  overlay.querySelector('#sun-start-clear')?.addEventListener('click', () => {
-    selected.clear();
-    slot.innerHTML = renderBodySilhouette(selected);
-    updateHint();
-  });
-  bindBodySilhouette(slot, selected, updateHint);
-  updateHint();
-
-  // Resolve the UVI lookup; render the pre-flight banner if conditions
-  // warrant it. Async — the dialog is already shown so we don't block.
-  uviPromise.then((uvi) => {
-    if (!Number.isFinite(uvi)) return;
-    latestPreflightUvi = uvi;
-    const banner = overlay.querySelector('#sun-start-uvi-banner');
-    if (!banner) return;
-    const html = _renderUVIPreflightBanner(uvi, fitz, psm);
-    if (html) {
-      banner.innerHTML = html;
-      banner.hidden = false;
-    }
-  }).catch(() => {});
-
-  overlay.querySelector('#start-confirm').addEventListener('click', async () => {
-    const eyeMode = overlay.querySelector('#start-eye-mode').value || 'direct';
-    const lensTint = overlay.querySelector('#start-lens-tint').value || 'clear';
-    const glassBetween = overlay.querySelector('#start-glass').checked;
-    const posture = overlay.querySelector('#start-posture').value || 'standing';
-    const surfaceAlbedo = overlay.querySelector('#start-surface').value || 'grass';
-    const rotatedSides = !!overlay.querySelector('#start-rotated')?.checked;
-    const regions = Array.from(selected);
-    if (regions.length === 0) {
-      hint.textContent = 'Tap at least one region before starting — what part of you is uncovered?';
-      hint.classList.add('sun-silhouette-hint-error');
-      setTimeout(() => hint.classList.remove('sun-silhouette-hint-error'), 2500);
-      return;
-    }
-    // Stash coords on the new session so the ticker can compute doses
-    // immediately without re-resolving location every tick.
-    const coords = getSunCoords();
-    const id = await startSession({ regions, eyeMode, lensTint, glassBetween, posture, surfaceAlbedo, rotatedSides, location: coords });
-    overlay.remove();
-    showNotification(_buildStartSessionToast({
-      regionCount: regions.length,
-      uvi: latestPreflightUvi,
-      psmTier: state.importedData?.sunDefaults?.photosensitiveMeds,
-      eyeMode,
-    }), 'success', 4500);
-    _refreshSurfaces();
-    _ensureActiveTicker();
-    return id;
-  });
-}
-
-// Focus management for dynamically-injected modals. Captures the current
-// focused element, lands focus on the first focusable inside the new
-// overlay, and restores focus to the trigger when the overlay is removed.
-// Wire backdrop click → close on a `.modal-overlay` element. Pairs the click
-// handler with a mousedown guard so a drag-from-inside-the-modal that
-// releases on the backdrop doesn't accidentally close (matches the global
-// _mouseDownInsideModal pattern in main.js for keyed overlays).
-//
-// Optional `closeFn` runs instead of plain `overlay.remove()` — needed for
-// modals with cleanup logic (camera streams in light-tools, focus restore,
-// etc.). Falls back to overlay.remove() when not given.
-export function _wireBackdropClose(overlay, closeFn) {
-  const close = typeof closeFn === 'function' ? closeFn : () => overlay.remove();
-  let mouseDownInside = false;
-  overlay.addEventListener('mousedown', (e) => {
-    mouseDownInside = !!e.target.closest('.modal');
-  });
-  overlay.addEventListener('click', (e) => {
-    if (mouseDownInside) { mouseDownInside = false; return; }
-    if (e.target === overlay) close();
-  });
-}
-
-// Single export so sun.js / views.js / light-tools.js share one helper.
-// Owns three concerns for any overlay-style modal:
-//   1. Auto-focus the first focusable element after first paint.
-//   2. Restore focus on teardown (any path — `.remove()`, parent rebuild,
-//      Escape close, backdrop click).
-//   3. Lock body scroll while the overlay is mounted so the page behind
-//      doesn't scroll under the modal on touch / wheel input.
-//   4. Bind Escape-to-close so keyboard users can dismiss without mousing.
-//
-// Stacks correctly with nested overlays by tracking the active overlay nodes:
-// only the outermost teardown restores `body.style.overflow`.
-const _modalScrollLocks = new Set();
-let _modalPriorOverflow = '';
-export function trapModalFocus(overlay) {
-  const previouslyFocused = document.activeElement;
-  // Body scroll lock: node tracking so stale teardown observers can't unlock
-  // the page while another overlay is still mounted.
-  if (_modalScrollLocks.size === 0) {
-    _modalPriorOverflow = document.body.style.overflow;
-  }
-  _modalScrollLocks.add(overlay);
-  document.body.style.overflow = 'hidden';
-  let teardown = false;
-  // Defer until after the browser paints — innerHTML may be set right
-  // after appendChild, and querySelector before paint can race.
-  setTimeout(() => {
-    const focusables = overlay.querySelectorAll(
-      'button:not([disabled]),input:not([disabled]),select:not([disabled]),textarea:not([disabled]),a[href],[tabindex]:not([tabindex="-1"])'
-    );
-    if (focusables.length > 0) try { focusables[0].focus(); } catch (e) {}
-  }, 30);
-  // Escape-to-close. Bound to document so it works regardless of where
-  // focus actually landed (e.g. focus traps that escaped the overlay).
-  const onKeydown = (e) => {
-    if (e.key === 'Escape' && document.body.contains(overlay)) {
-      e.preventDefault();
-      // Prefer overlay's own .remove() so any close() callback the modal
-      // wired (state cleanup, save-prompt, etc.) gets a chance to run via
-      // the MutationObserver below, rather than us bypassing it.
-      try { overlay.remove(); } catch (_) {}
-    }
-  };
-  document.addEventListener('keydown', onKeydown);
-  // Restore focus + release scroll-lock on overlay removal. MutationObserver
-  // catches every teardown path — .remove(), parent rebuild, Escape close.
-  const restore = () => {
-    if (teardown) return;
-    teardown = true;
-    document.removeEventListener('keydown', onKeydown);
-    _modalScrollLocks.delete(overlay);
-    if (_modalScrollLocks.size === 0) {
-      document.body.style.overflow = _modalPriorOverflow;
-    } else {
-      document.body.style.overflow = 'hidden';
-    }
-    if (previouslyFocused && typeof previouslyFocused.focus === 'function'
-        && document.contains(previouslyFocused)) {
-      try { previouslyFocused.focus(); } catch (e) {}
-    }
-  };
-  const obs = new MutationObserver(() => {
-    if (!document.body.contains(overlay)) {
-      obs.disconnect();
-      restore();
-    }
-  });
-  obs.observe(document.body, { childList: true, subtree: true });
-}
-
-// Plain-English session-stop summary. Leads with what the body got out of
-// the session (vit D in IU, top channel) and ends with the safety state —
-// the framing a normie reads after coming inside.
-function _plainStopSummary(sess, dur) {
-  if (!sess) return `Session saved — ${dur} min`;
-  const parts = [`Saved · ${dur} min outside`];
-  const fitz = sess.safety?.fitzpatrick || 'III';
-  const uvi = sess.atmosphere?.uvIndex;
-  const vitDAu = sess.doses?.vitamin_d || 0;
-  if (vitDAu > 0 && window.vitaminDIU) {
-    // Per-session cap (bodyFrac × 30k) when bodyFraction is known —
-    // avoids over-stating high-output device sessions whose raw IU
-    // would clamp at the 20k daily ceiling instead of the local
-    // skin-patch saturation (Audit P1 #8).
-    const bf = sess.bodyExposure?.fraction;
-    const iu = (Number.isFinite(bf) && bf > 0 && typeof window.vitaminDIUPerSession === 'function')
-      ? window.vitaminDIUPerSession(vitDAu, fitz, uvi, !!sess.bodyExposure?.rotatedSides, state.importedData?.genetics || null, bf)
-      : window.vitaminDIU(vitDAu, fitz, uvi, !!sess.bodyExposure?.rotatedSides);
-    if (iu >= 100) {
-      const lo = Math.round(iu * 0.6 / 50) * 50;
-      const hi = Math.round(iu * 1.5 / 50) * 50;
-      parts.push(`~${lo}–${hi} IU vitamin D`);
-    }
-  } else if (sess.bodyExposure?.glassBetween) {
-    parts.push('no vitamin D — glass blocks UVB');
-  } else if (uvi != null && uvi < 2) {
-    parts.push(`no vitamin D — UVI too low (${uvi.toFixed(1)})`);
-  }
-  const med = sess.safety?.medFraction || 0;
-  if (med >= 1.0) {
-    parts.push('over your burn threshold — no more sun today');
-  } else if (med >= 0.7) {
-    parts.push(`burn dose ${Math.round(med * 100)}% — close to limit, ease up`);
-  } else if (med >= 0.3) {
-    parts.push(`burn dose ${Math.round(med * 100)}% — well within safe range`);
-  }
-  return parts.join(' · ');
-}
-
-// Identify the strongest channel a session contributed to (for notification copy)
-function _topChannel(sess) {
-  if (!sess?.doses) return null;
-  let bestKey = null, bestVal = 0;
-  for (const [k, v] of Object.entries(sess.doses)) {
-    if (Number.isFinite(v) && v > bestVal) { bestVal = v; bestKey = k; }
-  }
-  if (!bestKey) return null;
-  const meta = CHANNEL_DISPLAY[bestKey];
-  const t = channelTier(bestVal, bestKey);
-  if (t === 0) return null;
-  return { label: meta?.label || bestKey, tier: tierLabel(t) };
-}
-
 // Re-render dashboard sidebar + current view after a session change so the
 // Light Today strip + sidebar entry appear / update without a manual reload.
 //
@@ -1768,738 +1385,6 @@ function _refreshSurfaces(scrollAnchor) {
     if (window.navigate) try { window.navigate(view, navOpts); } catch (e) {}
     setTimeout(() => _resumeActiveTickerIfNeeded(), 100);
   }, 150);
-}
-
-// First-fire jargon explainer for in-session toasts. Returns a one-line
-// definition the first time the user sees a piece of acronym-heavy copy
-// ("MED", "ICNIRP"), then '' on every subsequent fire so the running
-// session toasts stay terse. Persists per-key in localStorage so the
-// explanation isn't repeated across reloads. The dictionary is small and
-// scoped to the toasts that actually use jargon — we don't preface every
-// notification.
-const _JARGON_DEFINITIONS = {
-  med: 'MED = the smallest UV dose that turns your skin slightly pink (Fitzpatrick-tuned). ',
-  icnirp: 'ICNIRP = the body that publishes safe daily UV exposure limits for the eye. ',
-};
-function _jargonPrefix(key) {
-  if (typeof localStorage === 'undefined') return '';
-  const def = _JARGON_DEFINITIONS[key];
-  if (!def) return '';
-  const flag = `gb_jargon_seen_${key}`;
-  try {
-    if (localStorage.getItem(flag)) return '';
-    localStorage.setItem(flag, '1');
-  } catch (e) { return ''; }
-  return def;
-}
-
-// ─── Live in-progress session ticker ───────────────────────────────────
-//
-// While a session is active we want the on-screen card to feel alive:
-//   • elapsed time ticks every second (mm:ss, h:mm:ss past 1hr)
-//   • channel doses accumulate visibly — each minute outside, the user sees
-//     vit-D / circadian / NIR fill in
-//   • a single shared setInterval drives every active-session card on the
-//     page (dashboard strip + Light & Sun list both update)
-//
-// Strategy: compute a per-minute dose rate ONCE at session start (via the
-// usual reconstructSpectrum + computeChannelDoses path on the session's
-// midpoint) and cache it in the module-scoped _liveState map (NOT on the
-// session object — that would persist runtime-only fields to localStorage
-// and CRDT). The ticker then just multiplies by elapsed minutes — no
-// per-tick spectral math.
-
-let _activeTicker = null;
-
-// Live-ticker per-session state (rate snapshot, atm, zenith, fitzpatrick,
-// MED helper). Held in-memory only — NOT persisted on the session object,
-// so saveImportedData() never serializes the heavy `atm` blob or function
-// refs into localStorage / Evolu CRDT. Cleared on session end / delete.
-const _liveState = new Map(); // session.id → { ratePerMin, sedPerMin, fitzpatrick, atm, zenith, snapshotAt, fractionOfMEDFn, pending }
-
-function _getLiveState(id) { return _liveState.get(id) || null; }
-function _setLiveState(id, patch) {
-  const cur = _liveState.get(id) || {};
-  _liveState.set(id, Object.assign(cur, patch));
-}
-function _clearLiveState(id) { _liveState.delete(id); }
-
-function _formatElapsed(ms) {
-  const totalSec = Math.max(0, Math.floor(ms / 1000));
-  const h = Math.floor(totalSec / 3600);
-  const m = Math.floor((totalSec % 3600) / 60);
-  const s = totalSec % 60;
-  const pad = (n) => String(n).padStart(2, '0');
-  if (h > 0) return `${h}:${pad(m)}:${pad(s)}`;
-  return `${m}:${pad(s)}`;
-}
-
-// Snapshot the per-minute channel rate for the active session.
-//
-// Sliced model: each snapshot defines a "rate slice" that applies from
-// snapshotAt to the next snapshot (or to session end). committedDoses
-// accumulates the contribution of all closed slices; the current slice's
-// contribution is computed live in _liveDosesFor and added on top.
-//
-// First snapshot of a session: snapshotAt = sess.startedAt so the slice
-// covers from session start (handles page reload — first snapshot after
-// reload covers from start).
-// Re-snapshot (committedDoses already exists): snapshotAt = Date.now(),
-// the previous slice was committed by _commitCurrentSlice() before the
-// caller cleared ratePerMin.
-//
-// NEVER mutates the session object — keeps the atm payload + function
-// refs out of localStorage / CRDT.
-async function _snapshotActiveRate(sess) {
-  const cur = _getLiveState(sess.id);
-  if (cur && cur.ratePerMin) return cur;
-  if (cur && cur.pending) return null;
-  _setLiveState(sess.id, { pending: true });
-  try {
-    const reconstructSpectrum = window.reconstructSpectrum;
-    const computeChannelDoses = window.computeChannelDoses;
-    const erythemalSED = window.erythemalSED;
-    const fractionOfMED = window.fractionOfMED;
-    const solarZenithAngle = window.solarZenithAngle;
-    const fetchAtmosphere = window.fetchAtmosphere;
-    if (!reconstructSpectrum || !computeChannelDoses || !solarZenithAngle || !fetchAtmosphere) return null;
-    const coords = sess.location || getSunCoords();
-    if (!coords) return null;
-    const now = new Date();
-    let atm = await fetchAtmosphere({ lat: coords.lat, lon: coords.lon, isoTime: now.toISOString() });
-    atm = _applyAtmOverrides(atm);
-    // Mid-session source-flip guard: if a re-snapshot returns a result
-    // whose primary source differs from the prior snapshot AND the
-    // confidence dropped AND the UVI delta is large, REJECT the new atm
-    // and reuse the prior one. Targets the "airplane-mode toggle"
-    // pattern where the live ticker briefly fails over from CAMS (0.95
-    // confidence, real aerosol/AOD) to Open-Meteo (0.65, GFS-approx)
-    // and back, surfacing as a transient UVI spike (the v1.6.7 sanity
-    // warning catches this passively; this guard prevents the spike
-    // from feeding the live rate at all). 25% delta threshold rejects
-    // genuine source-driven discontinuities; smaller drifts pass through.
-    const priorAtm = _getLiveState(sess.id)?.atm;
-    if (priorAtm && Number.isFinite(priorAtm.uvIndex) && Number.isFinite(atm?.uvIndex)) {
-      const primarySrc = (s) => String(s || '').split('+')[0];
-      const sourcesDiffer = primarySrc(priorAtm.source) !== primarySrc(atm.source);
-      const priorConf = priorAtm.confidence ?? 0.6;
-      const newConf = atm.confidence ?? 0.6;
-      const downgraded = newConf < priorConf - 0.15;
-      const uviDelta = Math.abs(atm.uvIndex - priorAtm.uvIndex);
-      const largeJump = priorAtm.uvIndex > 0 && uviDelta > priorAtm.uvIndex * 0.25;
-      if (sourcesDiffer && downgraded && largeJump) {
-        atm = { ...priorAtm, _sourceFlipBlocked: { from: priorAtm.source, to: atm.source, attemptedUvi: atm.uvIndex, at: Date.now() } };
-      }
-    }
-    const zenith = solarZenithAngle(now, coords.lat, coords.lon);
-    const spectrum = reconstructSpectrum({
-      zenithDeg: zenith,
-      ozoneDU: atm.ozoneDU ?? 300,
-      altitudeM: coords.altitudeM ?? 0,
-      cloudCover: (atm.cloudCover ?? 0) / 100,
-    });
-    const liveBodyModifiers = {
-      glassBetween: !!sess.bodyExposure?.glassBetween,
-      sunscreenSPF: sess.bodyExposure?.sunscreenSPF || 0,
-    };
-    const ratePerMin = computeChannelDoses({
-      spectrum,
-      durationMin: 1,
-      bodyExposureFraction: sess.bodyExposure?.fraction ?? 0,
-      eyeExposure: sess.eyeExposure,
-      bodyModifiers: liveBodyModifiers,
-    });
-    const sedPerMin = erythemalSED({
-      spectrum,
-      durationMin: 1,
-      bodyExposureFraction: sess.bodyExposure?.fraction ?? 0,
-      bodyModifiers: liveBodyModifiers,
-    });
-    const lcSkin = state.importedData?.lightCircadian?.skinType;
-    const lcRoman = lcSkin && (window._skinTypeToFitzpatrick ? window._skinTypeToFitzpatrick(lcSkin) : (lcSkin.match(/^(I{1,3}|IV|VI?)\b/) || [])[1]);
-    const fitzpatrick = state.importedData?.sunDefaults?.fitzpatrick || lcRoman || 'III';
-    const psmTier = _normalizePSMTier(state.importedData?.sunDefaults?.photosensitiveMeds);
-    const medScale = photosensitiveMedScale(psmTier);
-    // baselineZenith is sampled once per session and never overwritten —
-    // keeps the per-slice zenithScale denominator stable so cumulative
-    // doses don't jump every refresh cycle.
-    const existing = _getLiveState(sess.id) || {};
-    // First snapshot: slice begins at sess.startedAt so all elapsed time
-    // counts. Re-snapshot (committedDoses already populated by the
-    // commit step): slice begins now.
-    const isReSnapshot = !!existing.committedDoses;
-    const sliceStart = isReSnapshot ? Date.now() : sess.startedAt;
-    _setLiveState(sess.id, {
-      ratePerMin, sedPerMin, fitzpatrick, medScale, psmTier, atm, zenith,
-      baselineZenith: existing.baselineZenith ?? zenith,
-      snapshotAt: sliceStart,
-      committedDoses: existing.committedDoses || {},
-      committedSED: existing.committedSED || 0,
-      committedRetinalUV: existing.committedRetinalUV || 0,
-      fractionOfMEDFn: fractionOfMED,
-      pending: false,
-    });
-    return _getLiveState(sess.id);
-  } catch (e) {
-    if (window.console && console.warn) console.warn('snapshotActiveRate failed', e);
-    _setLiveState(sess.id, { pending: false });
-    return null;
-  }
-}
-
-// Compute the per-minute channel rate + erythemal SED rate for a session
-// at a specific instant. Pulls interpolated atm fields from the cached
-// atmosphere's hourly arrays (so values smoothly cross hour boundaries
-// instead of step-changing) and computes a fresh spectrum using the live
-// solar zenith. Returns { rate, sedPerMin } in dose-per-minute units, or
-// null if any required engine module isn't wired yet.
-//
-// This is the inner kernel used by Simpson integration in _liveDosesFor /
-// _commitCurrentSlice — replaces the previous single-rate-times-elapsed
-// approximation with proper sub-slice spectral resolution.
-function _rateAtInstant(sess, instantMs) {
-  const live = _getLiveState(sess?.id);
-  if (!live || !live.atm) return null;
-  const reconstructSpectrum = window.reconstructSpectrum;
-  const computeChannelDoses = window.computeChannelDoses;
-  const erythemalSED = window.erythemalSED;
-  const solarZenithAngle = window.solarZenithAngle;
-  const interpolateAtmosphere = window.interpolateAtmosphere;
-  if (!reconstructSpectrum || !computeChannelDoses || !erythemalSED || !solarZenithAngle) return null;
-
-  const coords = sess.location;
-  if (!coords) return null;
-  const when = new Date(instantMs);
-  const isoTime = when.toISOString();
-
-  // Interpolate atm fields between hourly buckets when arrays available;
-  // otherwise fall back to the snapshot's scalar values.
-  let atmAtT = live.atm;
-  if (interpolateAtmosphere) {
-    const interp = interpolateAtmosphere(live.atm, isoTime);
-    if (interp) {
-      atmAtT = {
-        ...live.atm,
-        uvIndex: interp.uvIndex ?? live.atm.uvIndex,
-        cloudCover: interp.cloudCover ?? live.atm.cloudCover,
-        temperatureC: interp.temperatureC ?? live.atm.temperatureC,
-      };
-    }
-  }
-  // Re-apply user overrides on top of interpolated values so manual UVI
-  // takes precedence over both forecast + interpolation.
-  atmAtT = _applyAtmOverrides(atmAtT);
-
-  // Surface-orientation + albedo boost on the effective body fraction.
-  // Posture: standing/sitting/lying-supine/lying-prone. Albedo: surfaces
-  // (sand 25%, water 25%, snow 80%) reflect UV onto the body — modeled
-  // as a +(albedo × 0.5) multiplier (rough — half the reflected light
-  // reaches the body geometry from below).
-  const baseFraction = sess.bodyExposure?.fraction ?? 0;
-  const postureMult = _POSTURE_MULTIPLIERS[sess.posture] ?? 1.0;
-  const albedoMult = 1 + (_SURFACE_ALBEDO[sess.surfaceAlbedo] ?? 0) * 0.5;
-  const effFraction = baseFraction * postureMult * albedoMult;
-
-  const zenith = solarZenithAngle(when, coords.lat, coords.lon);
-  const spectrum = reconstructSpectrum({
-    zenithDeg: zenith,
-    ozoneDU: atmAtT.ozoneDU ?? 300,
-    altitudeM: coords.altitudeM ?? 0,
-    cloudCover: (atmAtT.cloudCover ?? 0) / 100,
-    aod: atmAtT?.airQuality?.aod ?? null,
-  });
-  const bodyModifiers = {
-    glassBetween: !!sess.bodyExposure?.glassBetween,
-    sunscreenSPF: sess.bodyExposure?.sunscreenSPF || 0,
-  };
-  const rate = computeChannelDoses({
-    spectrum,
-    durationMin: 1,
-    bodyExposureFraction: effFraction,
-    eyeExposure: sess.eyeExposure,
-    bodyModifiers,
-  });
-  const sedPerMin = erythemalSED({
-    spectrum,
-    durationMin: 1,
-    bodyExposureFraction: effFraction,
-    bodyModifiers,
-  });
-  // Retinal UV per minute — only nonzero when eye mode is 'direct' AND
-  // the sun is above the ~5° elevation threshold (zenith ≤ 85°). Same
-  // gate retinalUVdose() applies to hydrated sessions; without it the
-  // live ticker accrues phantom J/m² before "UV-A on" (Bird-Riordan
-  // emits non-zero weighted UV at high zenith that doesn't physically
-  // reach the eye). Linear ramp 85° → 80° matches the firstUVA window.
-  let retinalUVPerMin = 0;
-  if (sess.eyeExposure?.mode === 'direct') {
-    const elev = 90 - zenith;
-    let gate = 1.0;
-    if (elev <= 5) gate = 0;
-    else if (elev < 10) gate = (elev - 5) / 5;
-    retinalUVPerMin = _retinalUVPerMin(spectrum) * gate;
-  }
-  return { rate, sedPerMin, retinalUVPerMin };
-}
-
-// Posture orientation multipliers on bodyExposureFraction. Lying-supine
-// makes the front of the body nearly horizontal at noon → near-full beam
-// reception (~1.4× baseline standing). Lying-prone same for back. Sitting
-// is between standing and lying. These are rough — proper modeling would
-// require per-region cosine weighting based on actual body geometry.
-//
-// Cite: Diffey 1991 ("Solar UV exposure of the body" Phys. Med. Biol.
-// 36:299) reports posture-weighted body-form factors of ~0.95 for
-// standing, ~0.85 for sitting, and ~1.3-1.5 for supine at solar noon.
-// Webb et al. 2011 (Br. J. Dermatol.) replicates the supine 1.4× boost
-// in field measurements. The AI verdict for "I sunbathed lying down on
-// the beach" rides on this multiplier — values are rounded conservatively.
-const _POSTURE_MULTIPLIERS = {
-  standing:    1.0,
-  sitting:     0.85,
-  'lying-supine': 1.4,
-  'lying-prone':  1.4,
-};
-
-// Surface albedo (UV reflectance). 0.25 = sand/water; 0.80 = fresh snow.
-// Source: WHO INTERSUN guidance + CIE 174:2006.
-const _SURFACE_ALBEDO = {
-  grass:    0.03,
-  concrete: 0.10,
-  sand:     0.25,
-  water:    0.25,
-  snow:     0.80,
-};
-
-// Helper: integrate UV-band irradiance to get J/m² per minute at the eye.
-// Mirrors retinalUVdose() math but returns a rate (per-minute) instead of
-// total. Used by Simpson integration in _rateAtInstant.
-function _retinalUVPerMin(spectrum) {
-  if (!spectrum) return 0;
-  const dlambda = 5;
-  let uv = 0;
-  for (let i = 0; i < spectrum.irradiance.length; i++) {
-    const nm = spectrum.wavelengths[i];
-    if (nm > 400) break;
-    uv += spectrum.irradiance[i] * dlambda;
-  }
-  return uv * 60; // per-minute (60 s)
-}
-
-// Simpson's 1/3 rule integration of channel doses + SED + retinal UV
-// across [a, b] using 3 sample points (start, midpoint, end). Second-order
-// accurate vs the previous midpoint approximation, captures sub-slice
-// spectral shifts at low sun (where pure cosine zenith scaling
-// underestimates UVB drop). Cost: 3 spectrum + dose computes per call
-// (~15K JS ops, negligible).
-function _integrateSlice(sess, startMs, endMs) {
-  const durationMin = Math.max(0, (endMs - startMs) / 60000);
-  if (durationMin <= 0) return { doses: {}, sed: 0, retinalUV: 0 };
-  const midMs = (startMs + endMs) / 2;
-  const r0 = _rateAtInstant(sess, startMs);
-  const r1 = _rateAtInstant(sess, midMs);
-  const r2 = _rateAtInstant(sess, endMs);
-  if (!r0 || !r1 || !r2) return { doses: {}, sed: 0, retinalUV: 0 };
-  // Simpson: ∫ ≈ (b - a) × (f(a) + 4f(m) + f(b)) / 6
-  const doses = {};
-  for (const k of Object.keys(r1.rate)) {
-    const a = r0.rate[k] ?? 0;
-    const m = r1.rate[k] ?? 0;
-    const b = r2.rate[k] ?? 0;
-    doses[k] = durationMin * (a + 4 * m + b) / 6;
-  }
-  const sed = durationMin * (r0.sedPerMin + 4 * r1.sedPerMin + r2.sedPerMin) / 6;
-  const retinalUV = durationMin * (r0.retinalUVPerMin + 4 * r1.retinalUVPerMin + r2.retinalUVPerMin) / 6;
-  return { doses, sed, retinalUV };
-}
-
-// Commit the current rate slice's contribution into committedDoses. Called
-// just before re-snapshotting so the user-visible cumulative dose stays
-// correct across rate changes (cloud cover shifts, hour rollover, etc.).
-// Uses Simpson integration for sub-slice accuracy.
-function _commitCurrentSlice(sess) {
-  const live = _getLiveState(sess?.id);
-  if (!live || !live.ratePerMin || !live.snapshotAt) return;
-  const sliceStart = live.snapshotAt;
-  const sliceEnd = Date.now();
-  if (sliceEnd <= sliceStart) return;
-  const { doses, sed, retinalUV } = _integrateSlice(sess, sliceStart, sliceEnd);
-  const committedDoses = { ...(live.committedDoses || {}) };
-  for (const [k, v] of Object.entries(doses)) {
-    committedDoses[k] = (committedDoses[k] || 0) + v;
-  }
-  const committedSED = (live.committedSED || 0) + sed;
-  const committedRetinalUV = (live.committedRetinalUV || 0) + retinalUV;
-  _setLiveState(sess.id, { committedDoses, committedSED, committedRetinalUV });
-}
-
-// Compute live doses = committedDoses (sum of past, fully-closed slices) +
-// current-slice contribution integrated via Simpson's rule.
-//
-// Each Simpson sample reconstructs the spectrum at its instant using the
-// live solar zenith and INTERPOLATED atmosphere (linearly between the two
-// hourly forecast buckets surrounding the sample). This captures both
-// solar-angle drift AND sub-hourly atmospheric variation properly per
-// channel — vs the previous cosine-scaled-rate which underestimated UVB
-// attenuation at low sun (where Air Mass climbs non-linearly).
-//
-// Cost: 3 spectrum + dose computes per call. _liveDosesFor is invoked
-// from several places per tick — if profiling shows hotspots we can add
-// per-tick memoization keyed by tickCount.
-function _liveDosesFor(sess) {
-  const live = _getLiveState(sess?.id);
-  if (!live) return null;
-  // Paused sessions: surface committed totals only — current slice
-  // contributes zero. Skips the Simpson integration entirely.
-  if (sess?.paused) {
-    const committed = live.committedDoses || {};
-    const sed = live.committedSED || 0;
-    const retinalUV = live.committedRetinalUV || 0;
-    const medFraction = live.fractionOfMEDFn ? live.fractionOfMEDFn({ sed, fitzpatrick: live.fitzpatrick, medScale: live.medScale ?? 1.0 }) : 0;
-    return { doses: { ...committed }, sed, retinalUV, medFraction, fitzpatrick: live.fitzpatrick, psmTier: live.psmTier, atm: live.atm, paused: true };
-  }
-  if (!live.ratePerMin) return null;
-  const sliceStart = live.snapshotAt || sess.startedAt;
-  const now = Date.now();
-
-  const { doses: sliceDoses, sed: sliceSed, retinalUV: sliceRetinalUV } = _integrateSlice(sess, sliceStart, now);
-
-  const committed = live.committedDoses || {};
-  const doses = { ...committed };
-  for (const [k, v] of Object.entries(sliceDoses)) {
-    doses[k] = (doses[k] || 0) + v;
-  }
-  const sed = (live.committedSED || 0) + sliceSed;
-  const retinalUV = (live.committedRetinalUV || 0) + sliceRetinalUV;
-  const medFraction = live.fractionOfMEDFn ? live.fractionOfMEDFn({ sed, fitzpatrick: live.fitzpatrick, medScale: live.medScale ?? 1.0 }) : 0;
-  return { doses, sed, retinalUV, medFraction, fitzpatrick: live.fitzpatrick, psmTier: live.psmTier, atm: live.atm };
-}
-
-// Render a compact live card body — elapsed time, burn-risk %, channel chips.
-function _renderActiveCardBody(sess) {
-  const elapsed = _formatElapsed(Date.now() - sess.startedAt);
-  const live = _liveDosesFor(sess);
-  let medStr = '';
-  if (live && Number.isFinite(live.medFraction)) {
-    const pct = Math.round(live.medFraction * 100);
-    let label = 'safe', cls = '';
-    if (live.medFraction >= 1) { label = 'over threshold'; cls = 'over'; }
-    else if (live.medFraction >= 0.7) { label = 'high'; cls = 'warn'; }
-    else if (live.medFraction >= 0.3) { label = 'moderate'; cls = ''; }
-    medStr = `<span class="sun-session-med ${cls}" title="Burn dose so far — ${pct}% of your burn threshold (Fitzpatrick ${escapeAttr(live.fitzpatrick)})">${pct}% burn dose · ${escapeHTML(label)}</span>`;
-  }
-  const channelChips = live?.doses ? renderChannelChips(live.doses, sess) : '';
-  // Surface a live IU readout for vitamin D — the most user-resonant
-  // unit in the channel set. Computed from the same channel-au integral
-  // the chips render, just translated through vitaminDIU(). Hidden when
-  // the rate is essentially zero (cloudy / low UVB / behind glass).
-  let vitaminDStr = '';
-  if (live?.doses?.vitamin_d > 0) {
-    const elapsedMin = Math.max(0, (Date.now() - sess.startedAt) / 60000);
-    const fitz = live.fitzpatrick || sess.safety?.fitzpatrick || 'III';
-    const uvi = live.atm?.uvIndex ?? sess.atmosphere?.uvIndex ?? null;
-    // Live ticker uses the central estimate (the chip's already small;
-    // a range there gets too noisy). Detail modal surfaces the band.
-    const rotated = !!sess.bodyExposure?.rotatedSides;
-    // Live ticker prefers per-session cap when bodyFraction is set,
-    // mirroring the detail-modal rendering (Audit P1 #8).
-    const bf = sess.bodyExposure?.fraction;
-    const iu = (Number.isFinite(bf) && bf > 0 && typeof window.vitaminDIUPerSession === 'function')
-      ? window.vitaminDIUPerSession(live.doses.vitamin_d, fitz, uvi, rotated, state.importedData?.genetics || null, bf)
-      : (window.vitaminDIU ? window.vitaminDIU(live.doses.vitamin_d, fitz, uvi, rotated) : live.doses.vitamin_d * 60 * (rotated ? 2 : 1));
-    const ratePerMin = elapsedMin > 0 ? iu / elapsedMin : 0;
-    if (iu >= 50) {
-      const iuLabel = iu >= 10000 ? '~' + (iu / 1000).toFixed(1).replace(/\.0$/, '') + 'k IU'
-        : iu >= 1000 ? '~' + Math.round(iu / 100) * 100 + ' IU'
-        : '~' + Math.round(iu / 10) * 10 + ' IU';
-      const rateLabel = ratePerMin >= 100 ? `${Math.round(ratePerMin / 10) * 10} IU/min` : `${Math.round(ratePerMin)} IU/min`;
-      vitaminDStr = `<span class="sun-session-vitd" title="Approximate vitamin D₃ synthesis so far (central estimate; ±50% band — see session detail). Saturates around 20k IU per Holick photoisomerization plateau.">☀ ~${iuLabel} vit D · ${rateLabel}</span>`;
-    }
-  }
-  // Heat-stress chip — temperatureC > 30 + elapsed > 30 min. Visual
-  // affordance for the same condition that fires the showNotification
-  // alert (so users who dismissed the toast still see the cue).
-  let heatStr = '';
-  const tempC = live?.atm?.temperatureC ?? null;
-  const elapsedMin = (Date.now() - sess.startedAt) / 60000;
-  if (Number.isFinite(tempC) && tempC > 30 && elapsedMin > 30) {
-    heatStr = `<span class="sun-session-heat" title="Ambient ${tempC.toFixed(0)}°C — heat-stress risk rises with duration. Drink water, take a 10-min shade break.">🌡 ${Math.round(tempC)}°C · take a break</span>`;
-  }
-
-  // Retinal-UV chip — only meaningful when eye mode is 'direct'. Shows
-  // current cumulative ACTINIC-weighted UV at the eye (matches ICNIRP
-  // S(λ) basis). Daily limit 30 J/m²; warn at 15 J/m².
-  let retinalStr = '';
-  if (sess.eyeExposure?.mode === 'direct' && Number.isFinite(live?.retinalUV) && live.retinalUV > 3) {
-    const ruv = live.retinalUV;
-    const ruvDisplay = ruv >= 10 ? Math.round(ruv) : ruv.toFixed(1);
-    const cls = ruv >= 15 ? ' warn' : '';
-    const label = ruv >= 30 ? 'at ICNIRP daily limit' : ruv >= 15 ? 'half the daily limit' : 'building';
-    retinalStr = `<span class="sun-session-retinal${cls}" title="Actinic-weighted UV at the eye (≈ICNIRP S(λ)). Daily limit 30 J/m²; photokeratitis appears above ~50 J/m². At ${ruvDisplay} J/m² you're ${label}.">👁 ${ruvDisplay} J/m² eye UV</span>`;
-  }
-
-  return { elapsed, medStr, vitaminDStr, channelChips, heatStr, retinalStr };
-}
-
-// Update every active-session card on the page. Cheap — only DOM patches
-// for the elements that exist; no full re-render. Every 5 seconds also
-// refreshes the page-level channel grid + dashboard strip so the live
-// accumulated doses propagate beyond the session card itself.
-//
-// _lastChannelRefreshAt is a wall-time gate so pause/resume of the
-// ticker doesn't desync the cadence — pre-2026-05-08 used a global mod
-// counter (_tickCount % 5) which counted ticks regardless of whether
-// they fired.
-let _tickCount = 0;
-let _lastChannelRefreshAt = 0;
-const RETINAL_ALERT_GRACE_MS = 10 * 60 * 1000;
-function _tickActiveCards() {
-  const sessions = getSessions().filter(s => !s.endedAt);
-  if (sessions.length === 0) {
-    if (_activeTicker) { clearInterval(_activeTicker); _activeTicker = null; }
-    return;
-  }
-  _tickCount++;
-  for (const sess of sessions) {
-    const live = _getLiveState(sess.id);
-    // Skip rate-related work entirely while paused — the slice is committed
-    // and we want zero dose accrual from pausedAt → resume time.
-    if (sess.paused) {
-      // Paused cards still tick for display state; refresh DOM below but
-      // bypass snapshot/refresh + alerts.
-    } else {
-      // Lazy snapshot of the rate (async — fires once per session, cached
-      // in module-scoped _liveState map, never written to the session record)
-      if ((!live || !live.ratePerMin) && (!live || !live.pending)) _snapshotActiveRate(sess);
-    }
-
-    // Refresh the cached atmosphere snapshot every 5 min so cloud cover
-    // and UVI drift get reflected in the live rate. Commit the current
-    // slice's accumulated dose first (so the cumulative readout doesn't
-    // jump when the new rate replaces the old), then clear ratePerMin to
-    // force the next tick to re-snapshot. baselineZenith + committedDoses
-    // are preserved across refreshes by _snapshotActiveRate.
-    if (live && live.ratePerMin && !live.pending && !sess.paused) {
-      const last = live.snapshotAt || 0;
-      if (Date.now() - last > 5 * 60 * 1000) {
-        _commitCurrentSlice(sess);
-        _setLiveState(sess.id, { ratePerMin: null });
-      }
-    }
-
-    // Fire once at 70% MED (warning) and 100% MED (stop). Dedup via _liveState flags.
-    const liveDoses = _liveDosesFor(sess);
-    if (liveDoses && Number.isFinite(liveDoses.medFraction)) {
-      const med = liveDoses.medFraction;
-      const cur = _getLiveState(sess.id) || {};
-      if (med >= 1.0 && !cur.alertedOver) {
-        _setLiveState(sess.id, { alertedOver: true });
-        showNotification(_jargonPrefix('med') + 'Burn threshold reached. Move to shade or cover up. Hydrate, no more direct sun today — damage from here is cumulative.', 'error', 10000);
-      } else if (med >= 0.7 && !cur.alerted70) {
-        _setLiveState(sess.id, { alerted70: true });
-        showNotification(_jargonPrefix('med') + '70% of your burn dose. Best move: head into shade for ~10 min, then decide. If you stay, watch for skin warmth or pinkness.', 'warning', 8000);
-      }
-    }
-
-    // Retinal-UV alerts — only fire when eye mode is 'direct' (eyes
-    // uncovered + open toward sky). Sunglass / closed-eyes / behind-glass
-    // sessions accumulate zero retinal UV. retinalUV is now actinic-
-    // weighted (≈ ICNIRP S(λ)); ICNIRP daily exposure limit is 30 J/m²
-    // actinic, photokeratitis appears above ~50 J/m². 15 J/m² used as a
-    // half-way warning so the user can still react.
-    if (liveDoses && Number.isFinite(liveDoses.retinalUV) && sess.eyeExposure?.mode === 'direct') {
-      const ruv = liveDoses.retinalUV;
-      const cur = _getLiveState(sess.id) || {};
-      const elapsedMs = Date.now() - sess.startedAt;
-      if (elapsedMs < RETINAL_ALERT_GRACE_MS) {
-        // Starting under high UVI used to stack eye-risk toasts immediately.
-        // The active card still shows the eye-UV chip; threshold toasts are
-        // reserved for later crossings after the user has had time to settle.
-        _setLiveState(sess.id, {
-          alertedRetinal500: cur.alertedRetinal500 || ruv >= 15,
-          alertedRetinalOver: cur.alertedRetinalOver || ruv >= 30,
-        });
-      } else if (ruv >= 30 && !cur.alertedRetinalOver) {
-        _setLiveState(sess.id, { alertedRetinalOver: true, alertedRetinal500: true });
-        showNotification('Eye UV is high. Put on UV-blocking sunglasses or take a shade break.', 'warning', 8000);
-      } else if (ruv >= 15 && !cur.alertedRetinal500) {
-        _setLiveState(sess.id, { alertedRetinal500: true });
-        showNotification('Eye UV is building. Sunglasses or look-down breaks are a good idea.', 'warning', 6500);
-      }
-    }
-
-    // Heat-stress chip alert — fires once at 30 min into a session when
-    // temperatureC > 30. Heat exhaustion risk rises faster with duration
-    // than UV burn at high ambient; UV alerts alone don't catch this.
-    const tempC = liveDoses?.atm?.temperatureC ?? null;
-    const elapsedMinNow = (Date.now() - sess.startedAt) / 60000;
-    if (Number.isFinite(tempC) && tempC > 30 && elapsedMinNow > 30) {
-      const cur = _getLiveState(sess.id) || {};
-      if (!cur.alertedHeat) {
-        _setLiveState(sess.id, { alertedHeat: true });
-        showNotification(`${tempC.toFixed(0)}°C ambient — drink water, take a 10-min shade break. Heat exhaustion ramps faster than UV burn at this temperature.`, 'warning', 8000);
-      }
-    }
-
-    // DOM patches are only meaningful when the user can actually see the
-    // result. Bail when the tab is hidden (browser will pause rAF anyway,
-    // but the 1s setInterval keeps firing) or when none of the views that
-    // host live-session UI are active. Heat/UV alerts above still ran —
-    // they're toast notifications that don't depend on visible cards.
-    if (document.hidden) continue;
-    if (state.currentView !== 'light'
-        && state.currentView !== 'dashboard'
-        && !document.querySelector('.modal-overlay [data-id], .modal-overlay [data-live-elapsed-for]')) {
-      continue;
-    }
-
-    // Update any "live elapsed" text node on the page — dashboard Light
-    // Today CTA uses [data-live-elapsed-for] so the timer ticks every
-    // second from anywhere in the app.
-    const elapsedFmt = _formatElapsed(Date.now() - sess.startedAt);
-    document.querySelectorAll(`[data-live-elapsed-for="${CSS.escape(sess.id)}"]`).forEach(el => {
-      el.textContent = elapsedFmt;
-    });
-
-    const cards = document.querySelectorAll(`[data-id="${CSS.escape(sess.id)}"]`);
-    if (!cards.length) continue;
-
-    const body = _renderActiveCardBody(sess);
-
-    // Patch a chip in place rather than `outerHTML = …` so the node
-    // identity survives the tick. outerHTML rebuild discards focus +
-    // any inflight tooltip/a11y state every second; that breaks
-    // keyboard nav on the active card and makes screen readers
-    // re-announce the chip every second. textContent + className +
-    // title patches keep the existing element and only touch the
-    // attributes that actually changed.
-    const patchChip = (el, html) => {
-      if (!html) { el.remove(); return; }
-      const tmpl = document.createElement('template');
-      tmpl.innerHTML = html.trim();
-      const fresh = tmpl.content.firstElementChild;
-      if (!fresh) return;
-      if (el.className !== fresh.className) el.className = fresh.className;
-      const newTitle = fresh.getAttribute('title') || '';
-      if (el.getAttribute('title') !== newTitle) el.setAttribute('title', newTitle);
-      const newText = fresh.textContent;
-      if (el.textContent !== newText) el.textContent = newText;
-    };
-    cards.forEach(card => {
-      const durEl = card.querySelector('.sun-session-duration');
-      if (durEl) durEl.textContent = body.elapsed;
-      const medEl = card.querySelector('.sun-session-med');
-      if (medEl) patchChip(medEl, body.medStr);
-      else if (body.medStr) {
-        // Insert med chip into the head row if it doesn't exist yet
-        const head = card.querySelector('.sun-session-head .sun-session-duration');
-        if (head) head.insertAdjacentHTML('afterend', body.medStr);
-      }
-      const vitdEl = card.querySelector('.sun-session-vitd');
-      if (vitdEl) patchChip(vitdEl, body.vitaminDStr);
-      else if (body.vitaminDStr) {
-        // Insert vit-D chip after med chip (or after duration if no med yet)
-        const after = card.querySelector('.sun-session-med') || card.querySelector('.sun-session-duration');
-        if (after) after.insertAdjacentHTML('afterend', body.vitaminDStr);
-      }
-      // Heat chip — replace if present, insert in head row if not.
-      const heatEl = card.querySelector('.sun-session-heat');
-      if (heatEl) patchChip(heatEl, body.heatStr);
-      else if (body.heatStr) {
-        const after = card.querySelector('.sun-session-vitd') || card.querySelector('.sun-session-med') || card.querySelector('.sun-session-duration');
-        if (after) after.insertAdjacentHTML('afterend', body.heatStr);
-      }
-      // Retinal-UV chip — same pattern.
-      const retinalEl = card.querySelector('.sun-session-retinal');
-      if (retinalEl) patchChip(retinalEl, body.retinalStr);
-      else if (body.retinalStr) {
-        const after = card.querySelector('.sun-session-heat') || card.querySelector('.sun-session-vitd') || card.querySelector('.sun-session-med') || card.querySelector('.sun-session-duration');
-        if (after) after.insertAdjacentHTML('afterend', body.retinalStr);
-      }
-      // Channel chips: outerHTML still appropriate here — the channel
-      // wrapper rebuilds its child chip nodes (different rendered chips
-      // when tier rankings change), and the user can't focus inside it.
-      const oldChips = card.querySelector('.sun-channel-chips');
-      if (oldChips) oldChips.outerHTML = body.channelChips || '';
-      else if (body.channelChips) card.insertAdjacentHTML('beforeend', body.channelChips);
-    });
-  }
-  // Every 5s, refresh the surrounding "Channels this week" grid + Light
-  // Today dashboard strip. They read rollingChannelTotals which now mixes
-  // in the live partial doses, so re-rendering them shows accumulated UV-D
-  // / circadian / NIR rising in real time. Wall-time gate: 5000ms since
-  // last refresh, robust to ticker pause/resume.
-  const now = Date.now();
-  if (now - _lastChannelRefreshAt >= 5000) {
-    _lastChannelRefreshAt = now;
-    _refreshLiveChannelSurfaces();
-  }
-}
-
-// Re-render the channel grid + dashboard strip without forcing a full
-// `navigate()` (that would tear down the active modal / setup card / etc).
-function _refreshLiveChannelSurfaces() {
-  // Light & Sun page: replace the channels-section innerHTML in place
-  if (state.currentView === 'light' && window.renderLightChannelsLive) {
-    try { window.renderLightChannelsLive(); } catch (e) {}
-  }
-  // Dashboard: redraw the Light Today strip in place — but ONLY when
-  // the rendered HTML actually changed. The strip ticker fires every
-  // 5s; without this guard, an unchanged render still triggered an
-  // innerHTML swap, which re-instantiated every child element and
-  // caused CSS transitions on the pills + AI chip to flicker visibly
-  // ("blinking") even though no value had changed.
-  if (state.currentView === 'dashboard' && window.renderLightTodayStrip) {
-    const strip = document.querySelector('.light-today-strip');
-    if (strip) {
-      const html = window.renderLightTodayStrip();
-      if (html) {
-        const wrap = document.createElement('div');
-        wrap.innerHTML = html;
-        const fresh = wrap.firstElementChild;
-        if (fresh) {
-          for (const attr of fresh.getAttributeNames()) {
-            const newVal = fresh.getAttribute(attr);
-            if (strip.getAttribute(attr) !== newVal) strip.setAttribute(attr, newVal);
-          }
-          // Only swap children when the HTML genuinely differs. Strict
-          // string compare is cheap relative to the layout cost of a
-          // full subtree rebuild.
-          const freshInner = fresh.innerHTML;
-          if (strip.innerHTML !== freshInner) strip.innerHTML = freshInner;
-        }
-      }
-    }
-  }
-}
-
-// Start the global ticker. Idempotent — safe to call from multiple places
-// (session start, session resume after page reload, etc.).
-function _ensureActiveTicker() {
-  if (_activeTicker) return;
-  // Tick once immediately to populate on first paint, then every second.
-  _tickActiveCards();
-  _activeTicker = setInterval(_tickActiveCards, 1000);
-}
-
-// Re-establish the ticker on page load if a session is already active. Wired
-// into _refreshSurfaces + module init so navigation never leaves us silent.
-function _resumeActiveTickerIfNeeded() {
-  if (getActiveSession()) _ensureActiveTicker();
-}
-
-// Resolve session coordinates from the user's profile, country fallback, or a
-// previously cached precise location upgrade. Browser geolocation is no longer
-// asked at session-stop time — that ask lives in Settings → Light & Sun as an
-// explicit "Use precise location" upgrade.
-async function _hydrateFromProfileCoords(id) {
-  const coords = getSunCoords();
-  if (!coords) return;
-  const sess = getSessions().find(s => s.id === id);
-  if (!sess) return;
-  sess.location = { lat: coords.lat, lon: coords.lon, altitudeM: 0, source: coords.source };
-  await saveImportedData();
-  await hydrateSession(id);
 }
 
 // Country band → centroid lat (0=tropical, 4=subarctic). Used as the lat
@@ -2589,6 +1474,25 @@ function _summarizeBodyExposure(sess) {
   return 'Body unset';
 }
 
+configureSunActiveSession({
+  getSessions,
+  getActiveSession,
+  startSession,
+  stopSession,
+  hydrateSession,
+  getSunCoords,
+  saveImportedData,
+  applyAtmOverrides: _applyAtmOverrides,
+  refreshSurfaces: _refreshSurfaces,
+  normalizePSMTier: _normalizePSMTier,
+  photosensitiveMedScale,
+  exposurePresets: EXPOSURE_PRESETS,
+  eyeModes: EYE_MODES,
+  lensTints: LENS_TINTS,
+  postureOptions: POSTURE_OPTIONS,
+  surfaceOptions: SURFACE_OPTIONS,
+});
+
 configureSunSessionUI({
   getSessions,
   deleteSession,
@@ -2619,9 +1523,7 @@ configureSunSessionUI({
 // is built from the previous profile's selection set; the active-card
 // ticker keeps running with the prior profile's session list).
 function _resetSunModuleState() {
-  if (_activeTicker) { clearInterval(_activeTicker); _activeTicker = null; }
-  _tickCount = 0;
-  _lastChannelRefreshAt = 0;
+  resetSunActiveSessionState();
   resetBodySilhouetteState();
   _hydrateInFlight.clear();
 }
