+++ service-worker.js
@@ -272,6 +272,7 @@ const APP_SHELL = [
   '/js/brand-assets.js',
   // Sun + light modules are statically reachable from the app shell.
   '/js/sun.js',
+  '/js/sun-active-session.js',
   '/js/sun-session-ui.js',
   '/js/sun-body-silhouette.js',
   '/js/sun-ai-analysis.js',
